// jce.Front/src/app/@core/data/models/products/catalogGood.ts
export class CatalogGood {

  goodId: number;
  catalogId: number;
  title: string;
  details: string;
  indexId: string;
  isBasicProduct: boolean;
  isDisplayedOnJce: boolean;
  pintelSheetId: number;
  price: number;
  productType: number;
  refPintel: string;
  universID: number;
  CreatedBy: string;
  UpdatedBy: string;
  dateMin: string;
  dateMax: string;
  clientProductAlias: string;
  isAddedManually: boolean;
  employeeParticipationMessage: string;

}